refactor: split host opkssh routes
